Settings are now stored in a modal
I put the dropdown menu in a modal that is triggered by a button press. It still loads the other modals same as before and none of the parameter and json code as been changed. The css is a little janky but should be easy to figure out once we get the rerun setting button up in the chart.